защита в controller sape

+++ controllers/sape.php
@@ -52,7 +52,7 @@ Class Controller_Sape Extends Controller_Base {
         $id_project = self::$filter->out('int',(empty($_POST['id_project']) ? $_GET['id_project'] : $_POST['id_project']));
 
 
-        //if(self::$authentication["auth"] === true && self::$status == "admin"){
+        if(self::$authentication["auth"] === true && self::$status == "admin"){
             if($name_project != ""){
                 self::$sape->project_update($id_project, $name_project);
 
@@ -72,7 +72,7 @@ Class Controller_Sape Extends Controller_Base {
                 $this->template->vars('id_project', $id_project);
                 $this->template->view('edit_razdel');
             }
-        //}
+        }
     }
 
 
@@ -82,7 +82,7 @@ Class Controller_Sape Extends Controller_Base {
         $name_project = self::$filter->out('string',(empty($_POST['name_project']) ? '' : $_POST['name_project']));
         $kolichestvo_urls = self::$filter->out('string',(empty($_POST['kolichestvo_urls']) ? '' : $_POST['kolichestvo_urls']));
 
-        //if(self::$authentication["auth"] === true && self::$status == "admin"){
+        if(self::$authentication["auth"] === true && self::$status == "admin"){
             if($name_project != ""){
                 $project_id = self::$sape->project_add($name_project);
                 self::$sape_model->add_Razdel($project_id, $name_project, date("Y-m-d H:i:s"), $kolichestvo_urls);
@@ -93,7 +93,7 @@ Class Controller_Sape Extends Controller_Base {
                 $this->template->vars('kolichestvo_urls', $kolichestvo_urls);
                 $this->template->view('add_razdel');
             }
-        //}
+        }
     }
 
 
@@ -101,11 +101,11 @@ Class Controller_Sape Extends Controller_Base {
     {
         $id_project = self::$filter->out('string',(empty($_GET['id_project']) ? '' : $_GET['id_project']));
 
-        //if(self::$authentication["auth"] === true && self::$status == "admin"){
+        if(self::$authentication["auth"] === true && self::$status == "admin"){
             self::$sape->project_delete($id_project);
             self::$sape_model->delete_Razdel($id_project);
             $this->index();
-        //}
+        }
     }
 
 
@@ -115,7 +115,7 @@ Class Controller_Sape Extends Controller_Base {
 
         $id_project = self::$filter->out('int',(empty($_GET['id_project']) ? '' : $_GET['id_project']));
 
-        //if(self::$authentication["auth"] === true && self::$status == "admin"){
+        if(self::$authentication["auth"] === true && self::$status == "admin"){
             if($id_project != ""){
                 $project = self::$sape->get_project($id_project);
                 $info = self::$sape->get_urls($id_project);
@@ -124,7 +124,7 @@ Class Controller_Sape Extends Controller_Base {
                 $this->template->vars('info', $info);
                 $this->template->view('show_project');
             }
-        //}
+        }
 
     }
 
@@ -134,7 +134,7 @@ Class Controller_Sape Extends Controller_Base {
         $urls = empty($_POST['urls']) ? '' : $_POST['urls'];
         $id_project = self::$filter->out('int',(empty($_GET['id_project']) ? '' : $_GET['id_project']));
 
-        //if(self::$authentication["auth"] === true && self::$status == "admin"){
+        if(self::$authentication["auth"] === true && self::$status == "admin"){
             // Отображаем сам проект
             $project = self::$sape->get_project($id_project);
             $this->template->vars('project', $project);
@@ -155,17 +155,21 @@ Class Controller_Sape Extends Controller_Base {
             else{
                 $this->template->view('add_urls');
             }
-        //}
+        }
     }
 
 
     function delete_url()
     {
-        $id_url = self::$filter->out('int',(empty($_GET['id_url']) ? '' : $_GET['id_url']));
-        self::$sape->url_delete($id_url);
+        if(self::$authentication["auth"] === true && self::$status == "admin"){
+            $id_url = self::$filter->out('int',(empty($_GET['id_url']) ? '' : $_GET['id_url']));
+            self::$sape->url_delete($id_url);
 
-        $id_project = self::$filter->out('int',(empty($_GET['id_project']) ? '' : $_GET['id_project']));
-        $this->show_project();
+            $id_project = self::$filter->out('int',(empty($_GET['id_project']) ? '' : $_GET['id_project']));
+            $this->show_project();
+        }
+
+
     }
